+++ tests/testthat/test-estimate_ite.R
@@ -78,15 +78,6 @@ test_that("ITE Estimated Correctly", {
   expect_true(class(ite) == "numeric")
   expect_true(length(ite) == length(dts_1$y))
 
-  ite_method <- "bcf"
-  ite <- estimate_ite(y = dts_1$y, z = dts_1$z, X = dts_1$X,
-                      ite_method,
-                      include_ps = include_ps,
-                      ps_method = ps_method,
-                      oreg_method = oreg_method)
-  expect_true(class(ite) == "numeric")
-  expect_true(length(ite) == length(dts_1$y))
-
   ite_method <- "tpoisson"
   ite <- estimate_ite(y = round(abs(dts_1$y) + 1), z = dts_1$z, X = dts_1$X,
                       ite_method,